refactor(gateway): extract config/runtime-resolution helpers from run.py (slice 12 of #54962)
